% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decide.R
\name{decide}
\alias{decide}
\title{Decide which digital data and response}
\usage{
decide(covar, response, data)
}
\arguments{
\item{covar}{A string of which covariates to use. Can be gamma ("y"), EM ("em"), lidar ("DEM")
or combinations of gamma and EM ("y+em"), gamma and lidar ("y+dem") and EM with lidar ("em+dem").
specifying "all", chooses all combinations gamma + EM + lidar.}

\item{response}{A character of the response to be predicted (e.g., "pH").}

\item{data}{The data to be used.}
}
\value{
A dataframe of the digital data and response chosen.
}
\description{
Choose which digital data to use as covariates and which soil property to predict.
Essentially, the function cleans the data so that it only contains the necessary
information to make a model from.
}
\author{
Trevan Flynn
}
