% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densDistro.R
\name{densDistro}
\alias{densDistro}
\title{Plot summary statistics for training and test sets}
\usage{
densDistro(training, testing, type = "density", columns, ...)
}
\arguments{
\item{training}{The cleaned data from the decide() function of the training set.}

\item{testing}{The cleaned data from the decide() function of the training set.}

\item{columns}{The columns in the dataframe to make into long format.}

\item{...}{Additional arguments to be passed to facet_wrap() such as number of rows or columns (e.g., nrow, ncol)}

\item{geom}{The type of plot either "density" (kernel density distribution) or "boxplot".}
}
\value{
a ggplot density plot wrapped around each variable with colors representing the training and test for each.
}
\description{
A function to make a density plot of training and testing datasets. Good for visually
depicting the distribution of covariates or the response.
}
\note{
If order in which plots appear for the variables is needed this must be done before running the function
for example df$type = factor(df$type, levels = c("gamma data", "EM data")) otherwise will be in alphabetic order.
}
\author{
Trevan Flynn
}
