% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krUncert.R
\name{krUncert}
\alias{krUncert}
\title{Calculate the prediction limit range from regression kriging}
\usage{
krUncert(preds, kriged, limit = 0.95)
}
\arguments{
\item{preds}{Is a stars object of the final predictions of the regression kriging model.}

\item{kriged}{Is a stars object of the kriged residuals}

\item{limit}{Is the confidence level to calculated the upper and lower limits from.}
}
\value{
A stars object containing the lower limits, prediction range, and upper limits.
}
\description{
Calculates the the prediction range limit as well as the lower and upper prediction
intervals using a specified confidence level.
}
\author{
Trevan Flynn
}
