% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picpCalc.R
\name{picpCalc}
\alias{picpCalc}
\title{Calculate the prediction interval coverage probability (PICP) of a model}
\usage{
picpCalc(data, response, pred)
}
\arguments{
\item{data}{Is a data frame containing the measured soil property and the predicted soil property.}

\item{response}{A vector of the measured soil data (e.g., val$MSA).}

\item{pred}{A vector of the measured soil data (e.g., val$.pred)}
}
\value{
A list with the first element being a plot of PICP to confidence intervals
with Lin's concordance correlation coefficient and the second being the results of the
PICP series and confidence levels.
}
\description{
This function calculates the PICP for a series of confidence intervals from the
predictions of a model. The PICP is a type of sensitivity analysis for model
uncertainties, which indicates how well the formulated confidence intervals were
determined. In other words, do the predictions fall within the confidence intervals.
Ideally, the PICP should fall on a 1:1 line with confidence intervals indicating that
all predictions fall within the decided limits.
}
\references{
Brendan P. Malone , Budiman Minasny , Alex B. McBratney (2017). "Using R for Digital Soil Mapping", Springer
}
\author{
Trevan Flynn
}
