% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumStat.R
\name{sumStat}
\alias{sumStat}
\title{calculate summary statistics for datasets}
\usage{
sumStat(data, columns)
}
\arguments{
\item{data}{The data to calculate the stats on.}

\item{columns}{A vector (numeric or string) of the columns to evaluate.}
}
\value{
A data frame with the stats in columns and properties in rows
}
\description{
A simple function to calculate the min, mean, median, max, standard deviation,
quantiles (25% and 75% quantiles), skewness, kurtosis and CV.
}
\author{
Trevan Flynn
}
