% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_mod.R
\name{what_mod}
\alias{what_mod}
\title{Show which models are available in regFit()}
\usage{
what_mod()
}
\value{
A data frame
}
\description{
Shows which models are available for regFit(), the tidymodel (parsnip) name to look up
engines to use and the abbreviation to use in the regFit() function. For engines available, use show_engines() from tidymodels.
Note - Will be update constantly and cubist is still experimental.
}
\author{
Trevan Flynn
}
